(WIP) Add a lookup task for plugin use
This commit shows a general idea for adding a task based lookup plugin. Generally if a specific value is needed from a resource or if you want a list of attributes for a given resource this task allows you to look that up.
